+++ src/test/java/org/cicirello/sequences/distance/SequenceDistanceTests.java
@@ -32,7 +32,7 @@ import org.cicirello.permutations.Permutation;
 /**
  * JUnit tests for the various classes that implement sequence distance metrics.
  */
-public class SequenceDistanceTests {
+public class SequenceDistanceTests extends InternalTestHelpersSequenceDistance {
 	
 	private final static double EPSILON = 1e-10;
 	
@@ -97,30 +97,9 @@ public class SequenceDistanceTests {
 		assertEquals(2, d.distance(toList(c1),toList(c3)), "end points differ");
 	}
 	
-	private ArrayList<String> toList(String[] a) {
-		ArrayList<String> list = new ArrayList<String>();
-		for (String s : a) list.add(s);
-		return list;
-	}
 	
-	private ArrayList<NonComparable> toList(NonComparable[] a) {
-		ArrayList<NonComparable> list = new ArrayList<NonComparable>();
-		for (NonComparable s : a) list.add(s);
-		return list;
-	}
 	
-	private static class NonComparable {
-		private int id;
-		public NonComparable(int me) {
-			id = me;
-		}
-		public boolean equals(Object other) {
-			return ((NonComparable)other).id == id;
-		}
-		public int hashCode() {
-			return id;
-		}
-	}
+	
 	
 	@Test
 	public void testExactMatchDistance() {
@@ -692,154 +671,6 @@ public class SequenceDistanceTests {
 		}
 		assertEquals(51, d.distance(c1,c2));
 		assertEquals(51, d.distance(toList(c1),toList(c2)));
-	}
-	
-	@Test
-	public void testLCSObjectSequences() {
-		LongestCommonSubsequenceDistance d = new LongestCommonSubsequenceDistance();
-		String[] s1 = {"a","b","a","c","a","d","a","e","a","f","a","h","a","i","a","j","a"};
-		String[] s2 = {"k","a","m","n","o","p","a","l","a","a","q","a","a","a","a"};
-		assertEquals(16, d.distance(s1,s2));
-		assertEquals(16, d.distance(toList(s1),toList(s2)));
-		String[] s3 = {"a","b","a","c","a","d","a","e","a","f","a","h","a","i","a","j","a","z","a","z","a","z","a","z","a","z","a"};
-		assertEquals(26, d.distance(s3,s2));
-		assertEquals(26, d.distance(toList(s3),toList(s2)));
-		
-		NonComparable[] c1 = new NonComparable[17];
-		for (int i = 0; i < 17; i++) {
-			c1[i] = (i%2)==0 ? new NonComparable(0) : new NonComparable(i);
-		}
-		NonComparable[] c2 = new NonComparable[16];
-		for (int i = 0; i < 16; i++) {
-			c2[i] = (i%2==0) ? new NonComparable(100 + i) : new NonComparable(0);
-		}
-		assertEquals(17, d.distance(c1,c2));
-		assertEquals(17, d.distance(toList(c1),toList(c2)));
-	}
-	
-	@Test
-	public void testLongestCommonSubsequenceDistance() {
-		LongestCommonSubsequenceDistance d = new LongestCommonSubsequenceDistance();
-		identicalSequences(d);
-		// lcs of next pair is 8... lengths 17 and 15
-		// distance should be 17+15-2*8 = 16
-		String s1 = "abacadaeafahaiaja";
-		String s2 = "kamnopalaaqaaaa";
-		assertEquals(16, d.distance(s1,s2));
-		String s3 = "abacadaeafahaiajazazazazaza";
-		// increasing length of s1 without changing lcs should increase distance to 26
-		assertEquals(26, d.distance(s3,s2));
-		char[] a1 = s1.toCharArray();
-		char[] a2 = s2.toCharArray();
-		assertEquals(16, d.distance(a1,a2));
-		char[] a3 = s3.toCharArray();
-		assertEquals(26, d.distance(a3,a2));
-		{
-			boolean[] b1 = {true, false, true, false, true, false, true, true, true, true, true, true};
-			boolean[] b2 = {false, false, true, true, true, false, false, false};
-			// lcs is 5... lengths are 12 and 8
-			assertEquals(10, d.distance(b1,b2));
-			assertEquals(10.0, d.distancef(b1,b2), EPSILON);
-			boolean[] b3 = {false, false, true, true, true, false, false, false, false, false, false};
-			assertEquals(13, d.distance(b1,b3));
-			assertEquals(13.0, d.distancef(b1,b3), EPSILON);
-		}
-		{ // int
-			int[] b1 = new int[a1.length];
-			int[] b2 = new int[a2.length];
-			for (int i = 0; i < b1.length; i++) {
-				b1[i] = a1[i];
-			}
-			for (int i = 0; i < b2.length; i++) {
-				b2[i] = a2[i];
-			}
-			assertEquals(16, d.distance(b1,b2));
-			int[] b3 = new int[a3.length];
-			for (int i = 0; i < b3.length; i++) {
-				b3[i] = a3[i];
-			}
-			assertEquals(26, d.distance(b3,b2));
-		}
-		{ // long
-			long[] b1 = new long[a1.length];
-			long[] b2 = new long[a2.length];
-			for (int i = 0; i < b1.length; i++) {
-				b1[i] = a1[i];
-			}
-			for (int i = 0; i < b2.length; i++) {
-				b2[i] = a2[i];
-			}
-			assertEquals(16, d.distance(b1,b2));
-			long[] b3 = new long[a3.length];
-			for (int i = 0; i < b3.length; i++) {
-				b3[i] = a3[i];
-			}
-			assertEquals(26, d.distance(b3,b2));
-		}
-		{ // short
-			short[] b1 = new short[a1.length];
-			short[] b2 = new short[a2.length];
-			for (int i = 0; i < b1.length; i++) {
-				b1[i] = (short)a1[i];
-			}
-			for (int i = 0; i < b2.length; i++) {
-				b2[i] = (short)a2[i];
-			}
-			assertEquals(16, d.distance(b1,b2));
-			short[] b3 = new short[a3.length];
-			for (int i = 0; i < b3.length; i++) {
-				b3[i] = (short)a3[i];
-			}
-			assertEquals(26, d.distance(b3,b2));
-		}
-		{ // byte
-			byte[] b1 = new byte[a1.length];
-			byte[] b2 = new byte[a2.length];
-			for (int i = 0; i < b1.length; i++) {
-				b1[i] = (byte)a1[i];
-			}
-			for (int i = 0; i < b2.length; i++) {
-				b2[i] = (byte)a2[i];
-			}
-			assertEquals(16, d.distance(b1,b2));
-			byte[] b3 = new byte[a3.length];
-			for (int i = 0; i < b3.length; i++) {
-				b3[i] = (byte)a3[i];
-			}
-			assertEquals(26, d.distance(b3,b2));
-		}
-		{ // double
-			double[] b1 = new double[a1.length];
-			double[] b2 = new double[a2.length];
-			for (int i = 0; i < b1.length; i++) {
-				b1[i] = a1[i];
-			}
-			for (int i = 0; i < b2.length; i++) {
-				b2[i] = a2[i];
-			}
-			assertEquals(16, d.distance(b1,b2));
-			double[] b3 = new double[a3.length];
-			for (int i = 0; i < b3.length; i++) {
-				b3[i] = a3[i];
-			}
-			assertEquals(26, d.distance(b3,b2));
-		}
-		{ // float
-			float[] b1 = new float[a1.length];
-			float[] b2 = new float[a2.length];
-			for (int i = 0; i < b1.length; i++) {
-				b1[i] = a1[i];
-			}
-			for (int i = 0; i < b2.length; i++) {
-				b2[i] = a2[i];
-			}
-			assertEquals(16, d.distance(b1,b2));
-			float[] b3 = new float[a3.length];
-			for (int i = 0; i < b3.length; i++) {
-				b3[i] = a3[i];
-			}
-			assertEquals(26, d.distance(b3,b2));
-		}
 	}
 	
 	@Test
@@ -1533,39 +1364,7 @@ public class SequenceDistanceTests {
 	
 	
 	
-	private void identicalSequences(SequenceDistanceMeasurer d) {
-		for (int n = 0; n <= 10; n++) {
-			int[] a1 = new int[n];
-			long[] a2 = new long[n];
-			short[] a3 = new short[n];
-			byte[] a4 = new byte[n];
-			char[] a5 = new char[n];
-			float[] a6 = new float[n];
-			double[] a7 = new double[n];
-			boolean[] a8 = new boolean[n];
-			for (int i = 0; i < n; i++) {
-				a1[i] = ThreadLocalRandom.current().nextInt(100);
-				a2[i] = ThreadLocalRandom.current().nextInt(100);
-				a3[i] = (short)ThreadLocalRandom.current().nextInt(100);
-				a4[i] = (byte)ThreadLocalRandom.current().nextInt(100);
-				a5[i] = (char)ThreadLocalRandom.current().nextInt(100);
-				a6[i] = ThreadLocalRandom.current().nextInt(100);
-				a7[i] = ThreadLocalRandom.current().nextInt(100);
-				a8[i] = ThreadLocalRandom.current().nextBoolean();
-			}
-			String a9 = new String(a5);
-			// distance of a sequence to itself should be 0
-			assertEquals(0, d.distance(a1, a1.clone()));
-			assertEquals(0, d.distance(a2, a2.clone()));
-			assertEquals(0, d.distance(a3, a3.clone()));
-			assertEquals(0, d.distance(a4, a4.clone()));
-			assertEquals(0, d.distance(a5, a5.clone()));
-			assertEquals(0, d.distance(a6, a6.clone()));
-			assertEquals(0, d.distance(a7, a7.clone()));
-			assertEquals(0, d.distance(a8, a8.clone()));
-			assertEquals(0, d.distance(a9, new String(a9)));
-		}
-	}
+	
 	
 	private void identicalSequencesD(SequenceDistanceMeasurerDouble d) {
 		for (int n = 0; n <= 10; n++) {
